+++ internal/kwok-gpu-device-plugin/controllers/configmap/controller.go
@@ -1,4 +1,4 @@
-package configmamp
+package configmap
 
 import (
 	"log"